+++ main.cpp
@@ -1,6 +1,6 @@
 #include "raylib.h"
 #include "Animation.h"
-#include "GameLevel.h"
+#include "GameDesign.h"
 #include "Slashcode.h"
 #include <iostream>
 using namespace std;
@@ -95,6 +95,10 @@ int main()
             } break;
 
             case slashcode:{
+                      if (IsKeyPressed(KEY_ENTER) || IsGestureDetected(GESTURE_TAP))
+                {
+                    currentScreen = Ending;
+                }
 
               
             }break;
@@ -141,20 +145,8 @@ int main()
                     // TODO: Draw GAMEPLAY screen here!
                     //background
                     DrawRectangle(0, 0, screenWidth, screenHeight,GRAY);
-                    Color transparentBlack = Fade(BLACK, 0.8f);
-                    Rectangle roundedRect = { (screenWidth - 1002) / 2, (screenHeight - 983) / 2, 1002, 983 };
-                    DrawRectangleRounded(roundedRect, 0.1, 10, transparentBlack);
-                    
-                    DrawText("SLASHCODE", 509, 40, 70, WHITE);
-                    DrawRectangle(300,120, 850 , 3 , WHITE);
-                    DrawText("SELECT LEVEL", 559, 157, 40, WHITE);
-
-                    //button
-                    Rectangle buttonRect1 = { 344, 293, 764, 219 };
-                    Color buttonColor2 = BLACK;
-                    DrawRectangleRounded(buttonRect1, 0.4 , 10, buttonColor2);
-                    DrawText("BEGINNER LEVEL!", 378 , 368, 76 , WHITE);
                     settingsButton.Draw();
+                    settingsButton.DrawSetting();
                     
 
                    
@@ -166,48 +158,17 @@ int main()
                 {
                     // TODO: Draw GAMEPLAY screen here!
                     DrawRectangle(0, 0, screenWidth, screenHeight,GRAY);
-                    Color transparentBlack = Fade(BLACK, 0.8f);
-                    Rectangle roundedRect = { (screenWidth - 1002) / 2, (screenHeight - 983) / 2, 1002, 983 };
-                    DrawRectangleRounded(roundedRect, 0.1, 10, transparentBlack);
-                    
-                    DrawText("SLASHCODE", 509, 40, 70, WHITE);
-                    DrawRectangle(300,120, 850 , 3 , WHITE);
-                    DrawText("SELECT LANGUAGE", 520, 157, 40, WHITE);
-
-                    //button
-                    Rectangle buttonRect1 = { 344, 293, 764, 219 };
-                    Color buttonColor2 = BLACK;
-                    DrawRectangleRounded(buttonRect1, 0.4 , 10, buttonColor2);
-                    DrawText("PYTHON", 570 , 368, 76 , WHITE);
-                    settingsButton.Draw();
+                    settingsButton.Drawlanguage();
+                    settingsButton.DrawSetting();
                     
 
                 } break;
 
                 case slashcode:{
-                    DrawRectangle(0, 0, screenWidth, screenHeight,WHITE);
-                    DrawRectangle(0, 512, screenWidth, screenHeight,DARKGRAY);
-                    DrawRectangle(0, 512, screenWidth, 40 ,BLACK);
-                    DrawRectangle(400, 512, 5,512,BLACK );
-                    DrawRectangle(1040, 512, 5,512,BLACK );
-                    DrawRectangle(0,512,190,38,DARKGRAY);
-                    DrawRectangle(405,512,190,38,DARKGRAY);
-                    DrawRectangle(1050,512,190,38,DARKGRAY);
-                    DrawText("ACTIVITY",10 , 518, 30, WHITE);
-                    DrawText("Main.py",430 , 518, 30, WHITE);
-                    DrawText("OUTPUT",1070 , 518, 30, WHITE);
-                    DrawText("In Python, the print function is used to output ", 10, 565, 16, WHITE);
-                    DrawText("information to the console or terminal. It allows ", 10, 590, 16, WHITE);   
-                    DrawText("you to display text, variables, or expressions. ", 10, 610, 16, WHITE); 
-                    DrawText("The basic syntax of the print function is:", 10, 635, 16, WHITE); 
-                    DrawRectangle(5, 675, 380, 40,WHITE ); 
-                    DrawText("print(\"This is an example of printing a text.\")", 10, 685, 16, BLACK); 
-                    DrawRectangle(5, 775, 380, 100,WHITE );
-                    DrawText("INSTRUCTION:", 10, 790, 16, BLACK);  
-                    DrawText("print(\"Hello World!\")", 10, 830, 16, BLACK);               
-
+                    settingsButton.DrawSlashcode();
 
                 }break;
+
                 case Ending:
                 {
                     // TODO: Draw ENDING screen here!
@@ -222,6 +183,7 @@ int main()
     }
 
     // De-Initialization
+
     UnloadTexture(mySprite.spriteSheet);  // Unload texture
     CloseWindow();  // Close window and OpenGL context
